Refactor the Code Added more OOP.
