+++ src/main/java/Duke.java
@@ -7,199 +7,21 @@ import java.util.Scanner;
 public class Duke {
     public static void main(String[] args) {
         String inData = "yo";
-        String intro = "____________________________________________________________ \n"
-                + "Hello! I'm Joker \n"
-                + "What can I do for you? \n"
-                + "____________________________________________________________";
-        Task[] taskList = new Task[100];
-        int i = 0;
-        System.out.println(taskList[0]);
+        introMessage();
 
-        System.out.println(intro);
+        TaskManager taskManager = new TaskManager();
 
         Scanner scan = new Scanner(System.in);
         while (!Objects.equals(inData, "bye")) {
             inData = scan.nextLine();
-            if (Objects.equals(inData, "list")) {
-                System.out.println("____________________________________________________________\n" +
-                        "Here are the tasks in your list: ");
-                for (int j = 0; j < i; j++) {
-                    if (taskList[j] instanceof ToDo) {
-                        System.out.println(j + 1 + ". [" + taskList[j].getTask() + "]" +
-                                "[" + taskList[j].getStatusIcon() + "] " + taskList[j].getDescription());
-                    } else {
-                        System.out.println(j + 1 + ". [" + taskList[j].getTask() + "]" +
-                                "[" + taskList[j].getStatusIcon() + "] " + taskList[j].getDescription()
-                                + taskList[j].getDate());
-                    }
-                }
-                System.out.println("____________________________________________________________");
-            } else {
-                int inDataLength = inData.length();
-                if (inDataLength < 6) {
-                    try {
-                        if (!Objects.equals(inData, "bye") && Objects.equals(inData.substring(0, 4), "todo")) {
-                            inData.substring(0, 6);
-                        } else {
-                            try {
-                                if (!Objects.equals(inData, "bye") && !Objects.equals(inData.substring(0, 4), "done")) {
-                                    System.out.println("____________________________________________________________\n" +
-                                            "☹ OOPS!!! I'm sorry, but I don't know what that means :-(\n" +
-                                            "____________________________________________________________");
-                                }
-                            } catch (Exception e) {
-                                System.out.println("____________________________________________________________\n" +
-                                        "☹ OOPS!!! I'm sorry, but I don't know what that means :-(\n" +
-                                        "____________________________________________________________");
-                            }
-                        }
-                    } catch (Exception e) {
-                        System.out.println("____________________________________________________________\n" +
-                                "☹ OOPS!!! The description of a todo cannot be empty.\n" +
-                                "____________________________________________________________");
-                    }
-                } else {
-                    if (!inData.contains("deadline") && !inData.contains("todo") && !inData.contains("event")
-                            && !inData.contains("done") && !inData.contains("list") && !inData.contains("bye")
-                            && !inData.contains("delete")) {
-                        System.out.println("____________________________________________________________\n" +
-                                "☹ OOPS!!! I'm sorry, but I don't know what that means :-(\n" +
-                                "____________________________________________________________");
-                    } else {
-                        if (!Objects.equals(inData, "bye") && Objects.equals(inData.substring(0, 5), "todo ")) {
-                            taskList[i] = new ToDo(inData.substring(5, inDataLength));
-                            System.out.println(
-                                    "____________________________________________________________ \n"
-                                            + "Got it. I've added this task: \n"
-                                            + "[" + taskList[i].getTask() + "]"
-                                            + "[" + taskList[i].getStatusIcon() + "] "
-                                            + taskList[i].getDescription()
-                                            + "\n"
-                                            + "Now you have " + (i + 1) + " tasks in the list.\n"
-                                            + "____________________________________________________________");
-                            i++;
-
-                        }
-                        if (inData.contains("deadline") || inData.contains("event")) {
-                            if (Objects.equals(inData.substring(0, 9), "deadline ")) {
-                                String[] segments = inData.split("/by ");
-                                String date = segments[segments.length - 1];
-
-                                int segmentedLength = segments[0].length();
-                                String description = segments[0].substring(9, segmentedLength);
-                                taskList[i] = new Deadline(description, date);
-
-                                System.out.println(
-                                        "____________________________________________________________ \n"
-                                                + "Got it. I've added this task: \n"
-                                                + "[" + taskList[i].getTask() + "]"
-                                                + "[" + taskList[i].getStatusIcon() + "] "
-                                                + taskList[i].getDescription()
-                                                + "\n"
-                                                + "Now you have " + (i + 1) + " tasks in the list.\n"
-                                                + "____________________________________________________________");
-                                i++;
-
-                            } else if (Objects.equals(inData.substring(0, 6), "event ")) {
-                                String[] segments = inData.split("/at ");
-                                String date = segments[segments.length - 1];
-
-                                int segmentedLength = segments[0].length();
-                                String description = segments[0].substring(6, segmentedLength);
-                                taskList[i] = new Event(description, date);
-
-                                System.out.println(
-                                        "____________________________________________________________ \n"
-                                                + "Got it. I've added this task: \n"
-                                                + "[" + taskList[i].getTask() + "]"
-                                                + "[" + taskList[i].getStatusIcon() + "] "
-                                                + taskList[i].getDescription()
-                                                + "\n"
-                                                + "Now you have " + (i + 1) + " tasks in the list.\n"
-                                                + "____________________________________________________________");
-                                i++;
-                            }
-                        } else {
-                            if (Objects.equals(inData.substring(0, 5), "done ")) {
-                                if (isNumeric(inData.substring(5, inDataLength))) {
-                                    int taskNo = Integer.parseInt(inData.substring(5, inDataLength));
-                                    if (taskNo <= 100 && taskNo <= i) {
-                                        taskList[taskNo - 1].markAsDone();
-                                        System.out.println(
-                                                "____________________________________________________________\n"
-                                                        + "Nice! I've marked this task as done: \n"
-                                                        + " [" + taskList[taskNo - 1].getStatusIcon() + "] "
-                                                        + taskList[taskNo - 1].getDescription()
-                                                        + "\n____________________________________________________________");
-                                    } else {
-                                        System.out.println(
-                                                "____________________________________________________________ \n"
-                                                        + "invalid number!"
-                                                        + "\n____________________________________________________________");
-                                    }
-                                }
-                            }
-                            if (inData.contains("delete ")) {
-                                if (Objects.equals(inData.substring(0, 7), "delete ")) {
-                                    if (isNumeric(inData.substring(7, inDataLength))) {
-                                        int taskNo = Integer.parseInt(inData.substring(7, inDataLength));
-                                        if (taskNo <= 100 && taskNo <= i) {
-                                            System.out.println(
-                                                    "____________________________________________________________\n"
-                                                            + "Noted. I've removed this task: \n"
-                                                            + " [" + taskList[taskNo - 1].getStatusIcon() + "] "
-                                                            + taskList[taskNo - 1].getDescription()
-                                                            + "\n"
-                                                            + "Now you have " + (i - 1) + " tasks in the list.\n"
-                                                            + "____________________________________________________________");
-                                            taskList[taskNo - 1] = null;
-                                            if (taskNo < i + 1) {
-                                                for (int n = taskNo - 1; n < i; n++) {
-                                                    taskList[n] = taskList[n + 1];
-                                                }
-                                            }
-                                            i--;
-                                        }
-
-
-                                    }
-                                }
-                            }
-//                        } else {
-//                            jokerList[i] = inData;
-//                            taskList[i] = new Task(inData);
-//                            System.out.println(
-//                                    "____________________________________________________________ \n"
-//                                            + "added: "
-//                                            + inData
-//                                            + "\n"
-//                                            + "____________________________________________________________");
-//                            i++;
-//                        }
-                        }
-                    }
-                }
-            }
+            taskManager.executeCommand(inData);
         }
-        System.out.println(
-                "____________________________________________________________ \n"
-                            + "Bye. Hope to see you again soon! \n"
-                            + "____________________________________________________________");
     }
 
-    public static boolean isNumeric(String string) {
-        int intValue;
-
-        if(string == null || string.equals("")) {
-            System.out.println("String cannot be parsed, it is null or empty.");
-            return false;
-        }
-
-        try {
-            intValue = Integer.parseInt(string);
-            return true;
-        } catch (NumberFormatException e) {
-        }
-        return false;
+    public static void introMessage() {
+        System.out.println("____________________________________________________________ \n"
+                + "Hello! I'm Joker \n"
+                + "What can I do for you? \n"
+                + "____________________________________________________________");
     }
 }
